resize GL context on window size change
This is an API-breaking change for anyone not using the `Framework`
helper. Instead of calling the `set_frame_size()` method on the
`Renderer` after a window size change, it is now necessary to call the
`set_frame_size()` method on the `Engine`. `Framework` was updated to
handle this correctly.

Explicitly changing the GL context size (backbuffer) after window resize
events happens automatically on most platforms, but is needed for
Wayland and macOS according to glutin docs.

This fixes an issue on Wayland, where the window content size did not
change after a window resize, only the window title bar.

Might help for issues #167 and #10.