Fixes messages not being passed to plugins based on when they happened during frame
A 'frame' in fyrox can be oversimplified into the three parts that
I care about here:

- Use `user_interface.poll_message()` and pass results to plugins

..do some other stuff..

- Use `user_interface.poll_message()` again to clear any potential
  messages sent recently

- Draw frame

Now this is fine about half of the time, when a message is sent:

<-- here
First `poll_message()`

..do stuff..

Second `poll_message()`
<-- or here

But if a message is sent during the 'do stuff' part, since the
second `poll_message()` (in `rc/engine/executor.rs`) doesn't
actually do anything with the message recieved, only the widget(s)
involved recieve the message, and the plugins don't.

I found this problem in that my buttons were only sending
`on_ui_message` about half of the time, while still changing colour.